sane debug and adminRole
- debug only lives in ArkimeConfig, Config, WISESourceAPI now
- moved adminRole and check from Util to Auth, maybe YT is right
